feat(test): add FilterGroup test case
